Fixed Tool warnings
Resolved various warnings from the tools. CPD still complains about 1
duplication, which will be resolved when the two graph views are
combined. PMD doesn't have any warnings, and neither does Checkstyle.
FindBUgs still has 32 (out of 60 originally) bugs left, primarily
in the ExistingHandlerTest and RecentHandler(Test).

+++ src/test/java/toolbar/ExistingHandlerTest.java
@@ -13,19 +13,21 @@ import org.junit.Test;
  * @author Björn Ho
  */
 public class ExistingHandlerTest {
-	static String directory= System.getProperty("user.dir") 
+	static String directory = System.getProperty("user.dir") 
 			+ File.separator + "db" + File.separator;
 	
 	@BeforeClass
 	public static void cleanUp() {
-		for(File file: new File(directory).listFiles()) {
-			file.delete();
+		for (File file: new File(directory).listFiles()) {
+			if (file != null) {
+				file.delete();
+			}
 		}
 	}
 	
 	@Test
 	public void testMakeFileArray() throws IOException {
-		File file= new File(directory + "test.mv.db");
+		File file = new File(directory + "test.mv.db");
 		file.createNewFile();
 		ExistingHandler exHandler = new ExistingHandler();
 		assertEquals(directory + "test.mv.db", exHandler.makeFileArray()[0].toString());
